Refactor: Use corsFetch for all schemes
Currently corsFetch is very stable and does not need to be used differently from pyfetch depending on whether the license server values the Origin header.
Nested await is no longer necessary and the code is more concise.
It is expected to be less confusing for debugging and for people creating new schemes.