Update Readme
* Remove legacy version link from Readme
Remove the link from the Readme as there is little benefit for standard users to continue using the legacy version.
Legacy code will continue to be available from the Github branches list.

* Use appropriate vocabulary
The term "Content Decryption Module" should refer only to the binary module in which the function that performs the content decryption process and the Key pair are compiled together, and the term "CDM" in a broad sense for the key pair alone is not preferred.