Avoid editing libraries directly
Editing libraries directly is not a best practice.
This is because it can make it difficult for users or security auditors to know what has changed and can also may cause licensing issues.
Furthermore, the larger the change, the greater the effort required to update dependencies.

Instead, it is better to create appropriate overrides that allow you to change the result without modifying the original library's code.